Retrieve ViewModels with StoreOwner directly and not with the provider
