+++ ios/FantasyPremierLeague/FantasyPremierLeague/ViewModelStoreOwnerProvider.swift
@@ -26,29 +26,3 @@ struct ViewModelStoreOwnerProvider<Content: View>: View {
     }
 }
 
-/// A ViewModelStoreOwner specifically for iOS to be an ObservableObject.
-class IOSViewModelStoreOwner: ObservableObject, ViewModelStoreOwner {
-
-    var viewModelStore = ViewModelStore()
-    
-    /// This function allows retrieving the androidx ViewModel from the store.
-    func viewModel<T: ViewModel>(
-        key: String? = nil,
-        extras: CreationExtras? = nil
-    ) -> T {
-        do {
-            return try koinResolveViewModel(
-                vmClass: T.self,
-                owner: self,
-                key: key,
-                extras: extras
-            ) as! T
-        } catch {
-            fatalError("Failed to create ViewModel of type \(T.self)")
-        }
-    }
-    
-    func clear() {
-        viewModelStore.clear()
-    }
-}
